+++ crates/sdk-core/tests/integ_tests/workflow_tests/patches.rs
@@ -346,10 +346,7 @@ fn patch_marker_single_activity(
     let mut t = TestHistoryBuilder::default();
     t.add_by_type(EventType::WorkflowExecutionStarted);
     t.add_full_wf_task();
-    t.set_flags_first_wft(
-        &[CoreInternalFlags::UpsertSearchAttributeOnPatch as u32],
-        &[],
-    );
+    t.set_flags_first_wft(&[CoreInternalFlags::UpsertSearchAttributeOnPatch], &[]);
     match marker_type {
         MarkerType::Deprecated => {
             t.add_has_change_marker(MY_PATCH_ID, true);
@@ -720,10 +717,7 @@ async fn same_change_multiple_spots(#[case] have_marker_in_hist: bool, #[case] r
     let mut t = TestHistoryBuilder::default();
     t.add_by_type(EventType::WorkflowExecutionStarted);
     t.add_full_wf_task();
-    t.set_flags_first_wft(
-        &[CoreInternalFlags::UpsertSearchAttributeOnPatch as u32],
-        &[],
-    );
+    t.set_flags_first_wft(&[CoreInternalFlags::UpsertSearchAttributeOnPatch], &[]);
     if have_marker_in_hist {
         t.add_has_change_marker(MY_PATCH_ID, false);
         t.add_upsert_search_attrs_for_patch(&[MY_PATCH_ID.to_string()]);
@@ -830,10 +824,7 @@ async fn many_patches_combine_in_search_attrib_update(#[case] num_patches: usize
     let mut t = TestHistoryBuilder::default();
     t.add_by_type(EventType::WorkflowExecutionStarted);
     t.add_full_wf_task();
-    t.set_flags_first_wft(
-        &[CoreInternalFlags::UpsertSearchAttributeOnPatch as u32],
-        &[],
-    );
+    t.set_flags_first_wft(&[CoreInternalFlags::UpsertSearchAttributeOnPatch], &[]);
     for i in 1..=num_patches {
         let id = format!("patch-{i}");
         t.add_has_change_marker(&id, false);
